requirements.txt
boto3==1.17.97
botocore==1.20.97
dataclasses==0.6
Flask==1.1.4
jmespath==0.10.0
MarkupSafe==2.0.1
python-dateutil==2.8.1
s3transfer==0.4.2
six==1.16.0
urllib3==1.26.5
Werkzeug==1.0.1
python-jose==3.3.0